+++ compiler/src/org.graalvm.compiler.truffle.runtime/src/org/graalvm/compiler/truffle/runtime/EngineData.java
@@ -38,6 +38,7 @@ import static org.graalvm.compiler.truffle.options.PolyglotCompilerOptions.Compi
 import static org.graalvm.compiler.truffle.options.PolyglotCompilerOptions.CompileOnly;
 import static org.graalvm.compiler.truffle.options.PolyglotCompilerOptions.FirstTierCompilationThreshold;
 import static org.graalvm.compiler.truffle.options.PolyglotCompilerOptions.FirstTierMinInvokeThreshold;
+import static org.graalvm.compiler.truffle.options.PolyglotCompilerOptions.Inlining;
 import static org.graalvm.compiler.truffle.options.PolyglotCompilerOptions.MinInvokeThreshold;
 import static org.graalvm.compiler.truffle.options.PolyglotCompilerOptions.Mode;
 import static org.graalvm.compiler.truffle.options.PolyglotCompilerOptions.MultiTier;
@@ -106,6 +107,9 @@ public final class EngineData {
     @CompilationFinal public double splittingGrowthLimit;
     @CompilationFinal public int splittingMaxNumberOfSplitNodes;
 
+    // inlining options
+    @CompilationFinal public boolean inlining;
+
     // compilation options
     @CompilationFinal public boolean compilation;
     @CompilationFinal public boolean compileImmediately;
@@ -129,7 +133,6 @@ public final class EngineData {
 
     EngineData(OptionValues options) {
         this.id = engineCounter.incrementAndGet();
-        // splitting options
         loadOptions(options);
 
         // the reporter requires options to be initialized
@@ -138,6 +141,8 @@ public final class EngineData {
 
     void loadOptions(OptionValues options) {
         this.engineOptions = options;
+
+        // splitting options
         this.splitting = getPolyglotOptionValue(options, Splitting) &&
                         getPolyglotOptionValue(options, Mode) != EngineModeEnum.LATENCY;
         this.splittingAllowForcedSplits = getPolyglotOptionValue(options, SplittingAllowForcedSplits);
@@ -148,6 +153,10 @@ public final class EngineData {
         this.traceSplittingSummary = getPolyglotOptionValue(options, TraceSplittingSummary);
         this.splittingGrowthLimit = getPolyglotOptionValue(options, SplittingGrowthLimit);
         this.splittingMaxNumberOfSplitNodes = getPolyglotOptionValue(options, SplittingMaxNumberOfSplitNodes);
+
+        // inlining options
+        this.inlining = getPolyglotOptionValue(options, Inlining) &&
+                        getPolyglotOptionValue(options, Mode) != EngineModeEnum.LATENCY;
 
         // compilation options
         this.compilation = getPolyglotOptionValue(options, Compilation);