+++ compiler/src/org.graalvm.compiler.truffle.test/src/org/graalvm/compiler/truffle/test/TestWithSynchronousCompiling.java
@@ -45,7 +45,7 @@ import org.junit.Before;
  */
 public abstract class TestWithSynchronousCompiling extends TestWithPolyglotOptions {
 
-    private static final String[] DEFAULT_OTIONS = {
+    private static final String[] DEFAULT_OPTIONS = {
                     "engine.BackgroundCompilation", Boolean.FALSE.toString(), //
                     "engine.CompilationThreshold", "10", //
                     "engine.CompileImmediately", Boolean.FALSE.toString()
@@ -60,10 +60,10 @@ public abstract class TestWithSynchronousCompiling extends TestWithPolyglotOptio
     protected final Context setupContext(String... keyValuePairs) {
         String[] newOptions;
         if (keyValuePairs.length == 0) {
-            newOptions = DEFAULT_OTIONS;
+            newOptions = DEFAULT_OPTIONS;
         } else {
-            newOptions = Arrays.copyOf(DEFAULT_OTIONS, DEFAULT_OTIONS.length + keyValuePairs.length);
-            System.arraycopy(keyValuePairs, 0, newOptions, DEFAULT_OTIONS.length, keyValuePairs.length);
+            newOptions = Arrays.copyOf(DEFAULT_OPTIONS, DEFAULT_OPTIONS.length + keyValuePairs.length);
+            System.arraycopy(keyValuePairs, 0, newOptions, DEFAULT_OPTIONS.length, keyValuePairs.length);
         }
         return super.setupContext(newOptions);
     }